[SHA256 해싱 완료 및 세션 체크 완료]

+++ src/main/java/site/metacoding/white/web/BoardApiController.java
@@ -32,22 +32,25 @@ public class BoardApiController {
     private final BoardService boardService;
     private final HttpSession session;
 
-    @GetMapping("/v2/board/{id}")
-    public String findByIdV2(@PathVariable Long id) {
-        System.out.println("현재 open-in-view는 true 인가 false 인가 생각해보기!!");
-        boardService.findById(id);
-        // System.out.println("board.id : " + boardPS.getId());
-        // System.out.println("board.title : " + boardPS.getTitle());
-        // System.out.println("board.content : " + boardPS.getContent());
-        System.out.println("open-in-view가 false이면 Lazy 로딩 못함");
+    // @GetMapping("/v2/board/{id}")
+    // public String findByIdV2(@PathVariable Long id) {
+    // System.out.println("현재 open-in-view는 true 인가 false 인가 생각해보기!!");
+    // boardService.findById(id);
+    // // System.out.println("board.id : " + boardPS.getId());
+    // // System.out.println("board.title : " + boardPS.getTitle());
+    // // System.out.println("board.content : " + boardPS.getContent());
+    // System.out.println("open-in-view가 false이면 Lazy 로딩 못함");
 
-        // 날라감)
-        return "ok";
-    }
+    // // 날라감)
+    // return "ok";
+    // }
 
     @PostMapping("/board")
     public ResponseDto<?> save(@RequestBody BoardSaveReqDto boardSaveReqDto) {
         SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("로그인이 되지 않았습니다");
+        }
         // insert into board(title,content,user_id) values(?, ?, ?)
         boardSaveReqDto.setSessionUser(sessionUser);
         BoardSaveRespDto boardSaveRespDto = boardService.save(boardSaveReqDto); // 서비스에는 단 하나의 객체만 전달한다.
@@ -66,6 +69,10 @@ public class BoardApiController {
 
     @PutMapping("/board/{id}")
     public ResponseDto<?> update(@PathVariable Long id, @RequestBody BoardUpdateReqDto boardUpdateReqDto) {
+        SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("로그인이 되지 않았습니다");
+        }
         boardUpdateReqDto.setId(id);
         BoardUpdateRespDto boardPS = boardService.update(boardUpdateReqDto);
         return new ResponseDto<>(1, "성공", boardPS);
@@ -73,6 +80,10 @@ public class BoardApiController {
 
     @DeleteMapping("/board/{id}")
     public String deleteById(@PathVariable Long id) {
+        SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("로그인이 되지 않았습니다");
+        }
         boardService.deleteById(id);
         return "ok";
     }
