+++ src/main/java/site/metacoding/white/web/CommentApiController.java
@@ -24,13 +24,21 @@ public class CommentApiController {
 
     @PostMapping("/comment")
     public ResponseDto<?> save(@RequestBody CommentSaveReqDto commentSaveReqDto) {
+
         SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("로그인이 되지 않았습니다");
+        }
         commentSaveReqDto.setSessionUser(sessionUser);
         return new ResponseDto<>(1, "성공", commentService.save(commentSaveReqDto));
     }
 
     @DeleteMapping("/comment/{id}")
     public ResponseDto<?> deleteById(@PathVariable Long id) {
+        SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new RuntimeException("로그인이 되지 않았습니다");
+        }
         commentService.deleteById(id);
         return new ResponseDto<>(1, "성공", null);
     }
